fix: fix issues

+++ practice-two/src/javascript/modules/deleteNotes.js
@@ -1,10 +1,10 @@
-/* global document*/
+/* global document */
 import httpUtils from '../utils/httpUtils';
 import { renderAllNotes } from './getAndFilterNote';
 
 async function handleDeleteClick(noteId) {
   try {
-    await httpUtils.deleteNote(noteId); 
+    await httpUtils.deleteNote(noteId);
     await renderAllNotes();
     alert('Note deleted successfully!');
   } catch (error) {
@@ -12,15 +12,21 @@ async function handleDeleteClick(noteId) {
   }
 }
 
-const noteFullContainer = document.getElementById('note-full-container');
+document.addEventListener('DOMContentLoaded', () => {
+  const noteFullContainer = document.getElementById('note-full-container');
 
-noteFullContainer.addEventListener('click', function(event) {
-  if (event.target.classList.contains('remove-note')) {
-    const noteId = event.target.dataset.id;
-    const confirmDelete = confirm('Are you sure you want to delete this note?');
-    if (confirmDelete) {
-      handleDeleteClick(noteId);
-    }
+  if (noteFullContainer) {
+    noteFullContainer.addEventListener('click', function (event) {
+      if (event.target.classList.contains('remove-note')) {
+        const noteId = event.target.dataset.id;
+        const confirmDelete = window.confirm(
+          'Are you sure you want to delete this note?',
+        );
+        if (confirmDelete) {
+          handleDeleteClick(noteId);
+        }
+      }
+    });
   }
 });
